move window paq push into effect
