+++ app/react/V2/Components/Analitycs/specs/Matomo.spec.tsx
@@ -4,37 +4,53 @@
 
 import React from 'react';
 import { render } from '@testing-library/react';
+import { MemoryRouter } from 'react-router-dom';
 import { Provider } from 'jotai';
 import { atomsGlobalState } from 'V2/shared/testingHelpers';
 import { globalMatomoAtom, settingsAtom } from 'V2/atoms';
 import { Matomo } from '../Matomo';
 
-declare global {
-  interface Window {
-    _paq?: [string[]];
-  }
-}
-
 describe('Matomo', () => {
+  const originalLocation = window.location;
+
+  beforeAll(() => {
+    const mockLocation = new URL('https://mockedurl.com');
+    Object.defineProperty(window, 'location', {
+      value: mockLocation,
+      writable: true,
+    });
+  });
+
   beforeEach(() => {
     window._paq = undefined;
   });
+
+  afterAll(() => {
+    window.location = originalLocation;
+  });
+
+  const renderComponent = (store: any) => {
+    render(
+      <MemoryRouter>
+        <Provider store={store}>
+          <Matomo />
+        </Provider>
+      </MemoryRouter>
+    );
+  };
 
   it('should set the matomo config from the user config', () => {
     const store = atomsGlobalState();
     store.set(settingsAtom, { matomoConfig: '{"url":"https://url.org","id":"1"}' });
 
-    render(
-      <Provider store={store}>
-        <Matomo />
-      </Provider>
-    );
+    renderComponent(store);
 
     expect(window._paq).toStrictEqual([
-      ['trackPageView'],
-      ['enableLinkTracking'],
       ['setTrackerUrl', 'https://url.org/matomo.php'],
       ['setSiteId', '1'],
+      ['setCustomUrl', 'https://mockedurl.com/'],
+      ['trackPageView'],
+      ['enableLinkTracking'],
     ]);
   });
 
@@ -42,17 +58,14 @@ describe('Matomo', () => {
     const store = atomsGlobalState();
     store.set(globalMatomoAtom, { url: 'https://global.org', id: '1' });
 
-    render(
-      <Provider store={store}>
-        <Matomo />
-      </Provider>
-    );
+    renderComponent(store);
 
     expect(window._paq).toStrictEqual([
-      ['trackPageView'],
-      ['enableLinkTracking'],
       ['setTrackerUrl', 'https://global.org/tenant.php'],
       ['setSiteId', '1'],
+      ['setCustomUrl', 'https://mockedurl.com/'],
+      ['trackPageView'],
+      ['enableLinkTracking'],
     ]);
   });
 
@@ -63,18 +76,15 @@ describe('Matomo', () => {
     });
     store.set(globalMatomoAtom, { url: 'https://global.org', id: '2' });
 
-    render(
-      <Provider store={store}>
-        <Matomo />
-      </Provider>
-    );
+    renderComponent(store);
 
     expect(window._paq).toStrictEqual([
-      ['trackPageView'],
-      ['enableLinkTracking'],
       ['setTrackerUrl', 'https://global.org/tenant.php'],
       ['setSiteId', '2'],
       ['addTracker', 'https://url.org/matomo.php', '1'],
+      ['setCustomUrl', 'https://mockedurl.com/'],
+      ['trackPageView'],
+      ['enableLinkTracking'],
     ]);
   });
 
@@ -90,11 +100,7 @@ describe('Matomo', () => {
     store.set(settingsAtom, { matomoConfig: userJSON });
     store.set(globalMatomoAtom, { url: globalUrl, id: globalId });
 
-    render(
-      <Provider store={store}>
-        <Matomo />
-      </Provider>
-    );
+    renderComponent(store);
 
     expect(window._paq).toStrictEqual(undefined);
   });
@@ -105,19 +111,16 @@ describe('Matomo', () => {
     store.set(settingsAtom, { matomoConfig: '{"url":"https://url.org/","id":"10"}' });
     store.set(globalMatomoAtom, { url: 'https://global.org', id: '5' });
 
-    render(
-      <Provider store={store}>
-        <Matomo />
-      </Provider>
-    );
+    renderComponent(store);
 
     expect(window._paq).toStrictEqual([
       ['googleTracker', 'idForTracker'],
-      ['trackPageView'],
-      ['enableLinkTracking'],
       ['setTrackerUrl', 'https://global.org/tenant.php'],
       ['setSiteId', '5'],
       ['addTracker', 'https://url.org/matomo.php', '10'],
+      ['setCustomUrl', 'https://mockedurl.com/'],
+      ['trackPageView'],
+      ['enableLinkTracking'],
     ]);
   });
 
@@ -126,17 +129,14 @@ describe('Matomo', () => {
     store.set(settingsAtom, { matomoConfig: '{ malformed: "3",  }' });
     store.set(globalMatomoAtom, { url: 'https://global.org', id: '3' });
 
-    render(
-      <Provider store={store}>
-        <Matomo />
-      </Provider>
-    );
+    renderComponent(store);
 
     expect(window._paq).toStrictEqual([
-      ['trackPageView'],
-      ['enableLinkTracking'],
       ['setTrackerUrl', 'https://global.org/tenant.php'],
       ['setSiteId', '3'],
+      ['setCustomUrl', 'https://mockedurl.com/'],
+      ['trackPageView'],
+      ['enableLinkTracking'],
     ]);
   });
 });
